requirements.txt
azure-keyvault==1.1.0
fastapi
msrestazure
pymongo
pytest
uvicorn
aiofiles
passlib